refactor: split up and namespace store into multiple files
